bench: Refactor benchmark structure
- Move icosahedron implementation in a static library
- Add namespace
- Remove dead code
