+++ benchmark/CMakeLists.txt
@@ -12,4 +12,11 @@ function(add_coal_benchmark bench_name)
   target_include_directories(${target_name} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
 endfunction()
 
+add_library(${PROJECT_NAME}-bench-utils-icosahedron STATIC
+            utils/icosahedron.cpp)
+target_link_libraries(${PROJECT_NAME}-bench-utils-icosahedron
+                      PUBLIC Eigen3::Eigen)
+
 add_coal_benchmark(convex_support convex_support.cpp convex_support_highway.cpp)
+target_link_libraries(${PROJECT_NAME}-bench-convex_support
+                      PRIVATE ${PROJECT_NAME}-bench-utils-icosahedron)
